+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * Fallout 76 ([@TDarkShadow](https://www.github.com/TDarkShadow))
   * Fallout: London ([@TDarkShadow](https://www.github.com/TDarkShadow))
   * Warhammer 40,000: Darktide ([@LillyBird69](https://www.github.com/LillyBird69))
+  * Knights of the Old Republic II - The Sith Lords ([@JanuarySnow](https://www.github.com/JanuarySnow))
 
 #### Version - 4.0.1.0 - 4/18/2025
 * Fixed subfolders of profiles showing up under additional profiles within compiler settings